# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Watheq API's Docs",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "eng.asgamal@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/lawyer/login": {
            "post": {
                "tags": [
                    "Lawyer"
                ],
                "summary": "Check Lawyers then register or login them",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LawyerLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/LawyerLoginResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/lawyer/completeProfile": {
            "post": {
                "tags": [
                    "Lawyer"
                ],
                "summary": "Complete lawyer profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LawyerCompleteProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/LawyerLoginResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/lawyer/completeFiles": {
            "post": {
                "tags": [
                    "Lawyer"
                ],
                "summary": "Complete lawyer files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LawyerCompleteFilesRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/LawyerLoginResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/lawyer/updateVisibility": {
            "post": {
                "tags": [
                    "Lawyer"
                ],
                "summary": "Update lawyer visibility 1 fot online or 0 for offline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LawyerUpdateVisibilityRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/LawyerLoginResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/lawyer/updateProfile": {
            "post": {
                "tags": [
                    "Lawyer"
                ],
                "summary": "Update lawyer profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LawyerUpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/LawyerLoginResponses"
                        }
                    }
                }
            }
        },
        "/api/auth/lawyer/registerDeviceToken": {
            "post": {
                "tags": [
                    "Lawyer"
                ],
                "summary": "Register Lawyer device token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Device"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/auth/lawyer/logout": {
            "post": {
                "tags": [
                    "Lawyer"
                ],
                "summary": "Logout lawyer and remove device token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "X-Api-Language",
                        "in": "header",
                        "description": "['ar','en'] default is 'ar'",
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Logged in User access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Request body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RemoveDevice"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Device": {
            "properties": {
                "identifier": {
                    "type": "string",
                    "example": "D5bj4584545M48lOXq"
                },
                "firebaseToken": {
                    "type": "string",
                    "example": "F5PTn245O3aQ57"
                }
            },
            "type": "object",
            "xml": {
                "name": "Device"
            }
        },
...
        "LawyerLoginRequest": {
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "0532579547"
                }
            },
            "type": "object",
            "xml": {
                "name": "LawyerLoginRequest"
            }
        },
        "LawyerLoginResponses": {
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "name": {
                    "type": "string",
                    "example": "Ahmad"
                },
                "email": {
                    "type": "string",
                    "example": "ahmad@gmail.com"
                },
                "image": {
                    "type": "string",
                    "example": "uploads/ahmad.jpg"
                },
                "phone": {
                    "type": "string",
                    "example": "0532579547"
                },
                "created_at": {
                    "type": "integer",
                    "example": 1511649843
                },
                "isCompleteProfile": {
                    "type": "boolean",
                    "example": 1
                },
                "latitude": {
                    "type": "string",
                    "example": "40.7324319"
                },
                "longitude": {
                    "type": "string",
                    "example": "42.7324319"
                },
                "token": {
                    "type": "string",
                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOjIsImlzcyI6Imh0dHA6Ly8xMjcuMC4wLjE6ODAwMC9hcGkvY2xpZW50L2xvZ2luIiwiaWF0IjoxNTExNzI2MTkxLCJleHAiOjE1MTE3Mjk3OTEsIm5iZiI6MTUxMTcyNjE5MSwianRpIjoiSXFMck1NTU5sTlNFSVdBayJ9.B765ouFHZLcA1KyRVoDghO3zlvGey8glZ7wxIvL89To"
                }
            },
            "type": "object",
            "xml": {
                "name": "LawyerLoginResponses"
            }
        },
...
        "LawyerUpdateVisibilityRequest": {
            "properties": {
                "isOnline": {
                    "type": "boolean",
                    "enum": [
                        0,
                        1
                    ]
                }
            },
            "type": "object",
            "xml": {
                "name": "LawyerUpdateVisibilityRequest"
            }
        },
...
        "RemoveDevice": {
            "properties": {
                "identifier": {
                    "type": "string",
                    "example": "D5bj4584545M48lOXq"
                }
            },
            "type": "object",
            "xml": {
                "name": "RemoveDevice"
            }
        }
    },
...
}